# src/IntakeGate.Web/openapi/intake-gate.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "IntakeGate.Host | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18080/"
    }
  ],
  "paths": {
...
    "/api/ai/settings": {
      "get": {
        "tags": [
          "IntakeGate.Host"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSettingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/model-candidates/validate": {
      "post": {
        "tags": [
          "IntakeGate.Host"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateAiModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateAiModelResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/model-candidates/confirm": {
      "post": {
        "tags": [
          "IntakeGate.Host"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAiModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmAiModelResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AiSettingsResponse": {
        "required": [
          "profileConfigured",
          "provider",
          "model",
          "modelConfirmed",
          "ready",
          "modelValidatedAtUtc",
          "restartRequired",
          "activationMessage"
        ],
        "type": "object",
        "properties": {
          "profileConfigured": {
            "type": "boolean"
          },
          "provider": {
            "type": [
              "null",
              "string"
            ]
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "modelConfirmed": {
            "type": "boolean"
          },
          "ready": {
            "type": "boolean"
          },
          "modelValidatedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "restartRequired": {
            "type": "boolean"
          },
          "activationMessage": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ApiErrorResponse": {
        "required": [
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ConfirmAiModelRequest": {
        "required": [
          "confirmationToken"
        ],
        "type": "object",
        "properties": {
          "confirmationToken": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
      "ConfirmAiModelResponse": {
        "required": [
          "profileConfigured",
          "restartRequired",
          "activationMessage"
        ],
        "type": "object",
        "properties": {
          "profileConfigured": {
            "type": "boolean"
          },
          "restartRequired": {
            "type": "boolean"
          },
          "activationMessage": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ValidateAiModelRequest": {
        "required": [
          "provider",
          "model"
        ],
        "type": "object",
        "properties": {
          "provider": {
            "type": [
              "null",
              "string"
            ]
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
      "ValidateAiModelResponse": {
        "required": [
          "confirmationToken",
          "expiresAtUtc",
          "provider",
          "model",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "confirmationToken": {
            "type": "string"
          },
          "expiresAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}